additional restyling and renaming done so that the components make more sense for future coding, styling now proper for iphone 678
