+++ client/src/GlobalStyles.js
@@ -4,7 +4,6 @@ const GlobalStyles = createGlobalStyle`
 
 * {
   box-sizing: border-box;
-}
 
 body {
   margin: 0;
